+++ pkgs/unified_analytics/test/error_handler_test.dart
@@ -13,7 +13,7 @@ import 'package:unified_analytics/src/enums.dart';
 import 'package:unified_analytics/unified_analytics.dart';
 
 void main() {
-  late FileSystem fs;
+  late MemoryFileSystem fs;
   late Directory home;
   late FakeAnalytics initializationAnalytics;
   late FakeAnalytics analytics;
@@ -41,7 +41,7 @@ void main() {
 
     // This is the first analytics instance that will be used to demonstrate
     // that events will not be sent with the first run of analytics
-    initializationAnalytics = Analytics.test(
+    initializationAnalytics = Analytics.fake(
       tool: initialTool,
       homeDirectory: home,
       flutterChannel: flutterChannel,
@@ -51,7 +51,7 @@ void main() {
       dartVersion: dartVersion,
       fs: fs,
       platform: platform,
-    ) as FakeAnalytics;
+    );
     expect(initializationAnalytics.shouldShowMessage, true);
     initializationAnalytics.clientShowedMessage();
     expect(initializationAnalytics.shouldShowMessage, false);
@@ -61,7 +61,7 @@ void main() {
     //
     // This instance should have the same parameters as the one above for
     // [initializationAnalytics]
-    analytics = Analytics.test(
+    analytics = Analytics.fake(
       tool: initialTool,
       homeDirectory: home,
       flutterChannel: flutterChannel,
@@ -72,7 +72,7 @@ void main() {
       fs: fs,
       platform: platform,
       clientIde: clientIde,
-    ) as FakeAnalytics;
+    );
     analytics.clientShowedMessage();
 
     // The files that should have been generated that will be used for tests
@@ -93,7 +93,7 @@ void main() {
       expect(analytics.telemetryEnabled, false);
       expect(sessionFile.readAsStringSync(), isEmpty);
 
-      final secondAnalytics = Analytics.test(
+      final secondAnalytics = Analytics.fake(
         tool: initialTool,
         homeDirectory: home,
         flutterChannel: flutterChannel,
@@ -104,7 +104,7 @@ void main() {
         fs: fs,
         platform: platform,
         clientIde: clientIde,
-      ) as FakeAnalytics;
+      );
       expect(sessionFile.readAsStringSync(), isEmpty);
       expect(secondAnalytics.telemetryEnabled, false);