+++ pkgs/unified_analytics/test/legacy_analytics_test.dart
@@ -12,7 +12,7 @@ import 'package:unified_analytics/src/enums.dart';
 import 'package:unified_analytics/unified_analytics.dart';
 
 void main() {
-  late FileSystem fs;
+  late MemoryFileSystem fs;
   late Directory home;
   late Analytics analytics;
 
@@ -49,7 +49,7 @@ void main() {
 
     // The main analytics instance, other instances can be spawned within tests
     // to test how to instances running together work
-    analytics = Analytics.test(
+    analytics = Analytics.fake(
       tool: initialTool,
       homeDirectory: home,
       flutterChannel: flutterChannel,
@@ -80,7 +80,7 @@ void main() {
 
     // The main analytics instance, other instances can be spawned within tests
     // to test how to instances running together work
-    analytics = Analytics.test(
+    analytics = Analytics.fake(
       tool: initialTool,
       homeDirectory: home,
       flutterChannel: flutterChannel,
@@ -110,7 +110,7 @@ void main() {
 
     // The main analytics instance, other instances can be spawned within tests
     // to test how to instances running together work
-    analytics = Analytics.test(
+    analytics = Analytics.fake(
       tool: initialTool,
       homeDirectory: home,
       flutterChannel: flutterChannel,
@@ -140,7 +140,7 @@ void main() {
 
     // The main analytics instance, other instances can be spawned within tests
     // to test how to instances running together work
-    analytics = Analytics.test(
+    analytics = Analytics.fake(
       tool: initialTool,
       homeDirectory: home,
       flutterChannel: flutterChannel,
@@ -174,7 +174,7 @@ void main() {
 
     // The main analytics instance, other instances can be spawned within tests
     // to test how to instances running together work
-    analytics = Analytics.test(
+    analytics = Analytics.fake(
       tool: initialTool,
       homeDirectory: home,
       flutterChannel: flutterChannel,
@@ -208,7 +208,7 @@ void main() {
 
     // The main analytics instance, other instances can be spawned within tests
     // to test how to instances running together work
-    analytics = Analytics.test(
+    analytics = Analytics.fake(
       tool: initialTool,
       homeDirectory: home,
       flutterChannel: flutterChannel,
@@ -240,7 +240,7 @@ NOT VALID JSON
 
     // The main analytics instance, other instances can be spawned within tests
     // to test how to instances running together work
-    analytics = Analytics.test(
+    analytics = Analytics.fake(
       tool: initialTool,
       homeDirectory: home,
       flutterChannel: flutterChannel,
@@ -276,7 +276,7 @@ NOT VALID JSON
 
     // The main analytics instance, other instances can be spawned within tests
     // to test how to instances running together work
-    analytics = Analytics.test(
+    analytics = Analytics.fake(
       tool: initialTool,
       homeDirectory: home,
       flutterChannel: flutterChannel,
@@ -308,7 +308,7 @@ NOT VALID JSON
 
     // The main analytics instance, other instances can be spawned within tests
     // to test how to instances running together work
-    analytics = Analytics.test(
+    analytics = Analytics.fake(
       tool: initialTool,
       homeDirectory: home,
       flutterChannel: flutterChannel,